refactor existing DAO nav avatar to component
