+++ apps/admin/src/pages/Dao.tsx
@@ -3,19 +3,8 @@ import { DHLayout, useDHConnect } from '@daohaus/connect';
 import { useConnectedMember, useDao } from '@daohaus/moloch-v3-context';
 import { TXBuilder } from '@daohaus/tx-builder';
 import { useMemo } from 'react';
-import styled from 'styled-components';
-import { H4, ProfileAvatar } from '@daohaus/ui';
-
-const DaoNavElmContainer = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1.7rem;
-`;
-
-const DaoProfileAvatar = styled(ProfileAvatar)`
-  width: 4.8rem;
-  height: 4.8rem;
-`;
+import { H4 } from '@daohaus/ui';
+import { HeaderAvatar } from '../components/HeaderAvatar';
 
 export function Dao() {
   const { daochain, daoid } = useParams();
@@ -23,17 +12,6 @@ export function Dao() {
   const { provider, address } = useDHConnect();
   const { dao } = useDao();
   const { connectedMember } = useConnectedMember();
-
-  const daoNavElm = useMemo(() => {
-    if (dao) {
-      return (
-        <DaoNavElmContainer>
-          <DaoProfileAvatar image={dao.avatarImg} address={dao.id} />
-          <H4>{dao.name}</H4>
-        </DaoNavElmContainer>
-      );
-    }
-  }, [dao]);
 
   const moreLinks = useMemo(() => {
     if (connectedMember) {
@@ -66,7 +44,16 @@ export function Dao() {
       appState={{ dao }}
     >
       <DHLayout
-        leftNav={daoNavElm}
+        leftNav={
+          dao?.name &&
+          dao?.id && (
+            <HeaderAvatar
+              name={dao.name}
+              address={dao.id}
+              imgUrl={dao?.avatarImg}
+            />
+          )
+        }
         pathname={location.pathname}
         navLinks={[
           { label: 'Hub', href: `/${address}` },
